requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
waitress==3.0.2
beautifulsoup4==4.12.3
requests==2.31.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.2
cryptography==43.0.3
Flask-Migrate==4.0.5
python-dotenv==1.0.0
gunicorn==21.2.0
